Finish new button form
Added site & external reference fields and register button. Spec required SSID and Password fields as well, but commented these tests out. This is because we specified in the schema that these fields won't be required by the user. Instead we're using the external reference field.

Added create method to controller so form can complete registration and first test passes.